Refactor docker and scenario controller to helper function due to duplicating code
