+++ webserver/controllers/docker-controller.js
@@ -1,5 +1,12 @@
+const request = require('request');
 const dockerConnection = require('../config/docker-config');
 const util = require('../lib/utils');
+
+// URL Configuration for Master Server
+const masterProtocol = 'http://';
+const masterHost = process.env.MASTERHOST_PORT_1000_TCP_ADDR;
+const masterPort = 2000;
+const masterRoute = '/api/master';
 
 const status = {
   masterCount: 0,
@@ -19,4 +26,35 @@ const getMasterIP = (masterName, callback) => {
   });
 };
 
-module.exports = { createMaster, getMasterIP };
+const spoolUpMaster = data => {
+  if (process.env.NODE_ENV === 'production' || process.env.NODE_ENV === 'development') {
+    // create Master to execute new scenario
+    createMaster(
+      (masterName) => {
+        console.log('Master Name ==>', masterName);
+        // send data to master
+        setTimeout(() => {
+          getMasterIP(masterName, (masterIP) => {
+            console.log('Master IP received ==>', masterIP);
+            const masterUrl = `${masterProtocol}${masterIP}:${masterPort}${masterRoute}`;
+            console.log('Sending data to ==>', masterUrl);
+            data.masterName = masterName;
+            request.post({
+              url: masterUrl,
+              json: true,
+              body: data,
+            },
+            (err, response, body) => {
+              if (err) {
+                console.log('Error while sending data to master ==>', err);
+              } else {
+                console.log('Successfully sent data to master! Response body ==>', body);
+              }
+            });
+          });
+        }, 7000);
+      });
+  }
+};
+
+module.exports = { spoolUpMaster, createMaster, getMasterIP };
