+++ webserver/lib/utils.js
@@ -1,12 +1,3 @@
-const request = require('request');
-const dockerController = require('../controllers/docker-controller');
-
-// URL Configuration for Master Server
-const masterProtocol = 'http://';
-const masterHost = process.env.MASTERHOST_PORT_1000_TCP_ADDR;
-const masterPort = 2000;
-const masterRoute = '/api/master';
-
 const sendJSON = (res, status, content) => {
   res.status(status);
   res.json(content);
@@ -54,36 +45,4 @@ const checkContainer = (dockerConnection, containerName, callback) => {
   });
 };
 
-const spoolUpMaster = data => {
-  console.log('this should be called', data);
-  if (process.env.NODE_ENV === 'production' || process.env.NODE_ENV === 'development') {
-    // create Master to execute new scenario
-    dockerController.createMaster(
-      (masterName) => {
-        console.log('Master Name ==>', masterName);
-        // send data to master
-        setTimeout(() => {
-          dockerController.getMasterIP(masterName, (masterIP) => {
-            console.log('Master IP received ==>', masterIP);
-            const masterUrl = `${masterProtocol}${masterIP}:${masterPort}${masterRoute}`;
-            console.log('Sending data to ==>', masterUrl);
-            data.masterName = masterName;
-            request.post({
-              url: masterUrl,
-              json: true,
-              body: data,
-            },
-            (err, response, body) => {
-              if (err) {
-                console.log('Error while sending data to master ==>', err);
-              } else {
-                console.log('Successfully sent data to master! Response body ==>', body);
-              }
-            });
-          });
-        }, 7000);
-      });
-  }
-};
-
-module.exports = { sendJSON, createContainer, checkContainer, spoolUpMaster };
+module.exports = { sendJSON, createContainer, checkContainer };
